app initial commit

+++ src/tests/setup.ts
@@ -7,13 +7,15 @@ beforeAll(async () => {
   mongoServer = await MongoMemoryServer.create();
   const mongoUri = mongoServer.getUri();
   // Set the test URI in the global mongoose object
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   (global as any).mongoose = {
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     ...(global as any).mongoose,
     testUri: mongoUri,
     conn: null,
     promise: null
   };
-  process.env.NODE_ENV = 'test';
+  Object.defineProperty(process.env, 'NODE_ENV', { value: 'test' });
   await mongoose.connect(mongoUri);
 });
 
@@ -23,6 +25,9 @@ afterAll(async () => {
 });
 
 beforeEach(async () => {
+  if (!mongoose.connection.db) {
+    throw new Error('Database not connected');
+  }
   const collections = await mongoose.connection.db.collections();
   for (const collection of collections) {
     await collection.deleteMany({});
